backend/requirements.txt
Django>=4.2,<5
channels>=4.0
channels_redis>=4.0
djangorestframework>=3.14
psycopg2-binary>=2.9
celery>=5.3
django-celery-results>=2.5
redis>=4.5
django-storages[boto3]>=1.13
boto3>=1.26
python-dotenv>=1.0
weasyprint>=56.0
pywebpush>=1.13
qrcode>=7.3
djangorestframework-simplejwt>=5.2
stripe>=6.0
sendgrid>=6.0
django-cors-headers>=3.13
pytest>=7.0
pytest-django>=4.5
pytest-cov>=4.0
django-environ>=0.9
dj-database-url>=1.0
gunicorn>=20.1
moto[s3]>=4.0
responses>=0.22.0
